requirements.txt
type
psutil
matplotlib
numpy
importlib
pkgutil
